// backend/src/modulos/medalha/enums/tipo-medalha.enum.ts
export enum TipoMedalha {
  VELOCISTA = 'VELOCISTA', // Entregas rápidas (< 2min)
  MARATONISTA = 'MARATONISTA', // Alto volume de entregas
  PONTUAL = 'PONTUAL', // Alto SLA consistente
  MVP = 'MVP', // Top do ranking
  CONSISTENTE = 'CONSISTENTE', // Top 3 por período longo
  ROOKIE = 'ROOKIE', // Primeira entrega
}